requirements/base.txt


# Bleeding edge Django
django==1.9.6

# Configuration
django-environ==0.4.0
django-secure==1.0.1
whitenoise==3.0


# Forms
django-braces==1.8.1
django-crispy-forms==1.6.0
django-floppyforms==1.6.2

# Models
django-model-utils==2.5

# Images
Pillow==3.2.0

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.25.2


# Python-PostgreSQL Database Adapter
psycopg2==2.6.1

# Unicode slugification
unicode-slugify==0.1.3
django-autoslug==1.9.3

# Time zones support
pytz==2016.4

# Redis support
django-redis==4.4.2
redis>=2.10.0

# API
djangorestframework
